Refactor code for testing. Moved api tests from app to server.
