+++ server/package.json
@@ -44,6 +44,7 @@
   },
   "devDependencies": {
     "eslint": "^4.11.0",
+    "expect": "^21.2.1",
     "mocha": "^4.0.1",
     "request": "^2.83.0",
     "request-promise": "^4.2.2"
