+++ server/src/services/employees/employees.store.js
@@ -1,0 +1,12 @@
+const NeDB = require('nedb');
+const path = require('path');
+
+module.exports = function (app) {
+  const dbPath = app.get('nedb');
+  const Model = new NeDB({
+    filename: path.join(dbPath, 'payolicyChallenge.db'),
+    autoload: true
+  });
+
+  return Model;
+};
